+++ CHANGELOG.md
@@ -16,4 +16,5 @@
 - CMD+SHIFT+] on Mac / CMD+PGDN on Windows/Linux (next tab)
 - CMD+SHIFT+[ on Mac / CMD+PGUP on Windows/Linux (previous tab)
 - Use Chromium Engine
-- Ability to change engine on any platform through preferences
+- Ability to change engine through Preferences
+- Ability to change engine through Options Menu